BigDecimal

+++ src/main/java/com/gmail/ilerofv/entity/Report.java
@@ -5,11 +5,13 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @ToString
 @Setter
 @Getter
 @AllArgsConstructor
 public class Report {
     private int count;
-    private double average;
+    private BigDecimal average;
 }
